RHClient Documentation

Full implementation of Julia Client.

A client for Rest Harness, a robust tool that acts to stand in place of another RESTful service. This can be used to mock out a service that doesn't yet exist or where there is a need to reproduce a particular set of inputs from an opaque service. It can also serve as a simple shared cache. This software runs as a daemon service by default. This library contains a client developed in Julia that can be used to interact with Rest Harness.

Visit our Github page: RHClient.

RHClient.rhClient

RHClient.create_path

RHClient.create_paths

RHClient.delete_path

RHClient.delete_paths

RHClient.get_all

RHClient.get_path

RHClient.update_path

RHClient.rhClient — Type

rhClient(url::String)

Constructor for the REST Harness client

Arguments

url::String: concatenation of REST Harness IP + Port

RHClient.create_path — Method

create_path(c::rhClient, path::String, rc::Int64, return_value::String; 
    delay::Int64=0, headers::Any=nothing)

Send a "POST" HTTP request with a specified endpoint configuration to the REST Harness server.

Arguments

path::String: the name of the endpoint.

rc::Int64: the return code used by REST Harness when the endpoint is accessed.

return_value::String: the content of the endpoint.

delay::Int64=0: number of seconds REST Harness waits before returning a response to a "GET" request.

headers::Any=nothing: A dictionary containing headers to send with the "POST" request.

RHClient.create_paths — Method

create_paths(c::rhClient, path::Vector{Dict{String, Any}})

Send a "POST" HTTP request with a collection of endpoint configurations to the REST Harness server.

Arguments

path::Vector{Dict{String, Any}}: a collection of endpoint configuraions.

RHClient.delete_path — Method

delete_path(c::rhClient, path::String)

Send a "DELETE" HTTP request with an endpoint name specified to delete it from the REST Harness server.

Arguments

path::String: the name of the endpoint.

RHClient.delete_paths — Method

delete_paths(c::rhClient, path::String)

Send a "DELETE" HTTP request to delete all endpoint configurations from the REST Harness server.

Arguments

path::Vector{Dict{String, Any}}: a collection of endpoint configuraions.

RHClient.get_all — Method

get_all(c::rhClient)

Send a "GET" HTTP request to the REST Harness server and retrieve all of the currently stored endpoint configurations.

Return the endpoint configurations as a dictionary with path as the keys and a dictionary of the all other configurations as the values.

RHClient.get_path — Method

get_path(c::rhClient, path::String)

Send a "GET" HTTP request to the REST Harness server and retrieve the endpoint configuration.

Return endpoint configuration as a dictionary with path as the key and a dictionary of all other configurations as the value.

Arguments

path::String: the name of the endpoint.

RHClient.update_path — Method

update_path(c::rhClient, path::String, rc::Int64, return_value::String; 
    delay::Int64=0, headers=nothing)

Send a specified endpoint configuration to the REST Harness server to update an existing endpoint.

Arguments

path::String: the name of the endpoint.

rc::Int64: the return code used by REST Harness when the endpoint is accessed.

return_value::String: the content of the endpoint.

delay::Int64=0: number of seconds REST Harness waits before returning a response to a "GET" request.

headers::Any=nothing: A dictionary containing headers to send with the "POST" request.